services/worker/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
redis>=5.0.1
asyncpg>=0.28.0
sqlalchemy>=2.0.21
bullmq>=2.13.0
boto3>=1.34.0
charset-normalizer>=3.3.0
openai>=1.30.0
pypdf>=5.1.0
python-docx>=1.1.0
reportlab>=4.2.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
tiktoken>=0.7.0
